# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "cristinaperezboix.alu@iespacomolla.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/character/{character_id}/initial-event": {
            "post": {
                "tags": [
                    "Character"
                ],
                "summary": "Almacena un evento inicial para un personaje",
                "description": "Crea un registro de evento inicial en la base de datos para un personaje específico.",
                "operationId": "storeInitialEvent",
                "parameters": [
                    {
                        "name": "character_id",
                        "in": "path",
                        "description": "ID del personaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la página de inicio del dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Character created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personaje no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Personaje no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/character/{character_id}/ending-event": {
            "post": {
                "tags": [
                    "Character"
                ],
                "summary": "Almacena un evento de finalización para un personaje",
                "description": "Crea un registro de evento de finalización en la base de datos para un personaje específico.",
                "operationId": "storeEndingEvent",
                "parameters": [
                    {
                        "name": "character_id",
                        "in": "path",
                        "description": "ID del personaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la página de inicio del dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Game finished successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personaje no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Personaje no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/latest": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Obtener los últimos eventos",
                "description": "Obtiene los últimos eventos ordenados por fecha para cada personaje.",
                "operationId": "indexLatest",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}